update to new version and create LoggerUtils.ts
