+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "logger",
-  "version": "1.2.7",
+  "version": "1.2.8",
   "main": "./dist/index.js",
   "types": "./dist/types/index.d.ts",
   "engines": {
