+++ src/index.ts
@@ -8,11 +8,14 @@ import {
   IOReturnError,
   IOErrorParam,
   IOSetting,
+  IOErrorStack,
+  IOStd,
 } from "./LoggerInterfaces.js";
 import { LoggerConsole } from "./LoggerConsole.js";
 import { LoggerMethod } from "./LoggerMethod.js";
 import { LoggerProperty } from "./LoggerProperty.js";
 import { LoggerStatic } from "./LoggerStatic.js";
+import { LoggerUtils } from "./LoggerUtils.js";
 
 export default Logger;
 export {
@@ -20,6 +23,7 @@ export {
   LoggerStatic,
   LoggerMethod,
   LoggerProperty,
+  LoggerUtils,
   IOLoggerInterface,
   IOReturnGetTimeAndType,
   IOLevelLogId,
@@ -28,4 +32,6 @@ export {
   IOReturnError,
   IOErrorParam,
   IOSetting,
+  IOErrorStack,
+  IOStd,
 };
